now when connected to internet also the app loads fast and some mentor review comments were also covered
